+++ test/fixtures/r-package/R/helpers.R
@@ -1,0 +1,5 @@
+helper_a <- function(x) x + 1
+
+helper_b <- function(x, scale = 2) x * scale
+
+.internal <- function(x) x  # private, should be skipped