+++ test/fixtures/r-package/R/main.R
@@ -1,0 +1,6 @@
+source("R/helpers.R")
+
+run <- function(x, y) {
+  z <- helper_a(x)
+  siglocal::add(z, y)
+}